Added consumer system for when checking version
A consumer and a BiConsumer have been added to UpdateChecker.
The Consumer<Exception> is used to specify what to do if an exception is thrown during the cheking process. The default is to print the stack trace. This only runs if an exception is thrown.
The BiConsumer<CommandSender[], String> only runs if no exception is thrown and runs before the event. It allows to listen to the update checking process without having to create a listener. The first parameter are the requesters of the checking process, the second is the version readed from the spigot api